// qa-framework/src/library/models/Command.ts
enum Command {
  Visit = 'Visit',
  AssertEquals = 'Assert.Equals',
  Select = 'Select',
  Type = 'Type',
  Click = 'Click',
  WaitUntil = 'Wait.Until',
  Function = 'Function',
  Javascript = 'Javascript',
  MultiRun = 'MultiRun',
  IframeStart = 'IframeStart',
  IframeEnd = 'IframeEnd'
}

export default Command;